+++ frontend/src/components/auth/AuthRedirect.tsx
@@ -6,9 +6,9 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Loader2, Building, Users } from 'lucide-react';
-import { Card, CardContent, CardHeader, CardTitle } from '../components/ui/Card';
-import { Badge } from '../components/ui/badge';
-import { Button } from '../components/ui/Button';
+import { Card, CardContent, CardHeader, CardTitle } from '../../components/ui/Card';
+import { Badge } from '../../components/ui/badge';
+import { Button } from '../../components/ui/Button';
 
 interface UserContext {
   id: string;
