+++ frontend/src/components/auth/ForgotPassword.tsx
@@ -1,11 +1,11 @@
 // src/components/auth/ForgotPassword.tsx - Version moderne et optimisée
 import { useState, useEffect } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
-import { Button } from '../components/ui/Button';
-import { Card, CardContent, CardHeader, CardTitle } from '../components/ui/Card';
-import { Input } from '../components/ui/Input';
-import { Label } from '../components/ui/label';
-import { Alert, AlertDescription } from '../components/ui/alert';
+import { Button } from '../../components/ui/Button';
+import { Card, CardContent, CardHeader, CardTitle } from '../../components/ui/Card';
+import { Input } from '../../components/ui/Input';
+import { Label } from '../../components/ui/label';
+import { Alert, AlertDescription } from '../../components/ui/alert';
 import { useMultiTenantAuth } from '../../contexts/MultiTenantAuthContext';
 import { Loader2, Mail, ArrowLeft, CheckCircle, Info, AlertCircle, Shield, ArrowRight } from 'lucide-react';
