+++ frontend/src/components/auth/steps/OnboardingComplete.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Button } from '../components/ui/Button';
-import { Card, CardContent, CardHeader, CardTitle } from '../components/ui/Card';
+import { Button } from '../../ui/Button';
+import { Card, CardContent, CardHeader, CardTitle } from '../../ui/Card';
 import { CheckCircle, ArrowRight, Shield, Sparkles } from 'lucide-react';
 
 interface OnboardingCompleteProps {
