+++ frontend/src/components/layout/Footer.tsx
@@ -1,6 +1,6 @@
 // src/components/layout/Footer.tsx - Footer avec liens classiques
-import { Button } from "../components/ui/button";
-import { Separator } from "../components/ui/separator";
+import { Button } from "../ui/Button";
+import { Separator } from "../ui/separator";
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
