+++ frontend/src/components/lib/utils.ts
@@ -1,0 +1,2 @@
+export { cn } from '../../lib/utils';
+
