+++ frontend/src/components/manager/TeamScheduleView.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Card, CardContent, CardHeader, CardTitle } from '../components/ui/Card';
+import { Card, CardContent, CardHeader, CardTitle } from '../ui/Card';
 import { Calendar, Users } from 'lucide-react';
 
 interface TeamScheduleViewProps {
