+++ frontend/src/components/organization/MultiTenantDashboard.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import { AppLayout } from '../layout/AppLayout';
 import { useMultiTenantAuth, useTenant } from '../../contexts/MultiTenantAuthContext';
 import { ConditionalRender } from '../auth/ProtectedRoute';
-import { Card, CardContent, CardHeader, CardTitle } from '../components/ui/Card';
-import { Badge } from '../components/ui/badge';
+import { Card, CardContent, CardHeader, CardTitle } from '../ui/Card';
+import { Badge } from '../ui/badge';
 import { 
   Users, 
   Calendar,
