+++ frontend/src/components/qr/QRCodeScanner.tsx
@@ -1,10 +1,10 @@
 // src/components/qr/QRCodeScanner.tsx - Composant de scan de QR code
 
 import { useState, useRef, useEffect } from 'react';
-import { Card, CardContent, CardHeader, CardTitle } from '../components/ui/Card';
-import { Button } from '../components/ui/Button';
-import { Alert, AlertDescription } from '../components/ui/alert';
-import { Badge } from '../components/ui/badge';
+import { Card, CardContent, CardHeader, CardTitle } from '../ui/Card';
+import { Button } from '../ui/Button';
+import { Alert, AlertDescription } from '../ui/alert';
+import { Badge } from '../ui/badge';
 import { 
   Camera, 
   CameraOff,
