+++ frontend/src/pages/Integrations/IntegrationsDashboard.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { Card } from '../components/ui/Card';
-import { Button } from '../components/ui/Button';
-import { Badge } from '../components/ui/badge';
+import { Card } from '../../components/ui/Card';
+import { Button } from '../../components/ui/Button';
+import { Badge } from '../../components/ui/badge';
 import { 
   Zap, 
   Calendar,
