+++ frontend/src/pages/Users/UsersList.tsx
@@ -1,13 +1,13 @@
 // src/pages/Users/UsersList.tsx - Liste des utilisateurs avec gestion des rôles
 import { useState, useEffect } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
-import { useAuth, usePermissions } from '../hooks/use-auth';
-import { Card, CardContent, CardHeader, CardTitle } from '../components/ui/Card';
-import { Button } from '../components/ui/Button';
-import { Badge } from '../components/ui/badge';
-import { Input } from '../components/ui/input';
-import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../components/ui/select';
-import { Avatar, AvatarFallback, AvatarImage } from '../components/ui/avatar';
+import { useAuth, usePermissions } from '../../hooks/use-auth';
+import { Card, CardContent, CardHeader, CardTitle } from '../../components/ui/Card';
+import { Button } from '../../components/ui/Button';
+import { Badge } from '../../components/ui/badge';
+import { Input } from '../../components/ui/input';
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../../components/ui/select';
+import { Avatar, AvatarFallback, AvatarImage } from '../../components/ui/avatar';
 import { 
   Users, 
   Search, 
@@ -23,7 +23,7 @@ import {
   UserX,
   Shield
 } from 'lucide-react';
-import { userService } from '../services';
+import { userService } from '../../services';
 import type { User, UserRole, UserStatus } from '../../shared';
 import { toast } from 'react-toastify';
