+++ frontend/vite.config.ts
@@ -53,13 +53,7 @@ export default defineConfig({
   ],
   resolve: {
     alias: {
-      '@': path.resolve(__dirname, './src'),
-      '../components': path.resolve(__dirname, './src/components'),
-      '../pages': path.resolve(__dirname, './src/pages'),
-      '../services': path.resolve(__dirname, './src/services'),
-      '../hooks': path.resolve(__dirname, './src/hooks'),
-      '../context': path.resolve(__dirname, './src/context'),
-      '../utils': path.resolve(__dirname, './src/utils')
+      '@': path.resolve(__dirname, './src')
     }
   },
   server: {
